fix: update statements did not work properly
sqlalchemy is too smart,
it decides whether we need to commit changes
based on first line of raw sql statement.
https://stackoverflow.com/questions/45347565/executing-multiple-statements-with-postgresql-via-sqlalchemy-does-not-persist-ch
It was a script from Bohdan
https://github.com/near/near-indexer-for-explorer/pull/99#issuecomment-838871700
Bohdan's update started from `UPDATE` and everything worked.
Mine started from `WITH` and sqlalchemy decided not to commit anything because of that.
We lost some logging, but at least it works